* SysLinker
- Common_def.h

1. LOGIC 폴더 정의 추가

- Define.h

1. 프로젝트 번호 갱신 (1.5 -> 1.6)

2. 프로젝트 루트 폴더 정의 추가 (기존 C:\\Ficon3에서 C:\\SLP4 폴더로 루트 폴더 변경)

- HelpMsgManager

1. 입력 회로 트리에서 채널만 존재하고 회로가 없을 경우 툴팁 기능에서 말단 노드로 판단해서 설비명 가져오다가 죽는 오류 수정

- NewExcelManager

1. 프로젝트 루트 폴더 변경으로 인한 수정 (실질적으로는 사용되지 않는 매서드들이 대부분이지만 수정함)

- MainFrm

1. 프로젝트 루트 폴더 변경으로 인한 수정

- SysLinker

1. 프로젝트 루트 폴더 변경으로 인한 수정

*** 그 외의 추가 및 변경 사항은 [UI 기반 펌프 편집]  기능에 관련된 것으로 김형섭 부장 추가

+++ SysLinker/FormRelayEdit.cpp
@@ -276,7 +276,7 @@ LRESULT CFormRelayEdit::OnEditDlgPreview(WPARAM wp, LPARAM lp)
 	else if (lp == DATA_SAVE)
 		MakeChangeDeviceMap(m_pNewData, CHANGE_ACTION_UPDATE);
 	if (m_pDlgBottomChangePane)
-		m_pDlgBottomChangePane->EditDlgPreviewChange(wp,lp);
+		m_pDlgBottomChangePane->EditDlgPreviewChange(wp,(int)lp);
 	return 0;
 }
